+++ src/main/java/com/akash/quizzy/service/QuestionService.java
@@ -21,7 +21,7 @@ public class QuestionService {
         return questionDao.findByCategory(category);
     }
 
-    public Question getQuestionById(Integer id) {
+    public Question getQuestionById(Long id) {
         return questionDao.findById(id).get();
     }
